refactor: enhance type usage and import structure in ColorableMergedBodyMaterial and ColorableMergedEdgeMaterial
